// processors/aftermath/src/types/sui/0xc4049b2d1cc0f6e017fda8260e4377cecd236bd7f56a54fee120816e72e2e0dd.ts
    export const TYPE_QNAME =
      "0xefe170ec0be4d762196bedecd7a065816576198a6527c99282a2551aaa7da38c::admin::AdminCap";
// ...
    export const TYPE_QNAME =
      "0xefe170ec0be4d762196bedecd7a065816576198a6527c99282a2551aaa7da38c::admin::AuthCap";
// ...
export class events extends SuiBaseProcessor {
  constructor(options: SuiBindOptions) {
    super("events", options);
  }
  static DEFAULT_OPTIONS: SuiBindOptions = {
    address:
      "0xefe170ec0be4d762196bedecd7a065816576198a6527c99282a2551aaa7da38c",
    network: SuiNetwork.MAIN_NET,
  };

  static bind(options: Partial<SuiBindOptions> = {}): events {
    return new events({ ...events.DEFAULT_OPTIONS, ...options });
  }

  onEventCreatedPoolEvent(
    func: (event: events.CreatedPoolEventInstance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): events {
    this.onMoveEvent(func, { type: "events::CreatedPoolEvent" }, fetchConfig);
    return this;
  }

  onEventDepositEvent(
    func: (event: events.DepositEventInstance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): events {
    this.onMoveEvent(func, { type: "events::DepositEvent" }, fetchConfig);
    return this;
  }

  onEventDepositEventV2(
    func: (event: events.DepositEventV2Instance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): events {
    this.onMoveEvent(func, { type: "events::DepositEventV2" }, fetchConfig);
    return this;
  }

  onEventOraclePriceEvent(
    func: (event: events.OraclePriceEventInstance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): events {
    this.onMoveEvent(func, { type: "events::OraclePriceEvent" }, fetchConfig);
    return this;
  }

  onEventSpotPriceEvent(
    func: (event: events.SpotPriceEventInstance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): events {
    this.onMoveEvent(func, { type: "events::SpotPriceEvent" }, fetchConfig);
    return this;
  }

  onEventSwapEvent(
    func: (event: events.SwapEventInstance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): events {
    this.onMoveEvent(func, { type: "events::SwapEvent" }, fetchConfig);
    return this;
  }

  onEventSwapEventV2(
    func: (event: events.SwapEventV2Instance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): events {
    this.onMoveEvent(func, { type: "events::SwapEventV2" }, fetchConfig);
    return this;
  }

  onEventWithdrawEvent(
    func: (event: events.WithdrawEventInstance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): events {
    this.onMoveEvent(func, { type: "events::WithdrawEvent" }, fetchConfig);
    return this;
  }

  onEventWithdrawEventV2(
    func: (event: events.WithdrawEventV2Instance, ctx: SuiContext) => void,
    fetchConfig?: Partial<MoveFetchConfig>,
  ): events {
    this.onMoveEvent(func, { type: "events::WithdrawEventV2" }, fetchConfig);
    return this;
  }
}

export namespace events {
  export interface CreatedPoolEvent {
    pool_id: _0x2.object$.ID;
    name: string;
    creator: string;
    lp_type: _0x1.ascii.String;
    coins: _0x1.ascii.String[];
    weights: bigint[];
    flatness: bigint;
    fees_swap_in: bigint[];
    fees_swap_out: bigint[];
    fees_deposit: bigint[];
    fees_withdraw: bigint[];
  }

  export namespace CreatedPoolEvent {
    export const TYPE_QNAME =
      "0xefe170ec0be4d762196bedecd7a065816576198a6527c99282a2551aaa7da38c::events::CreatedPoolEvent";

    const TYPE = new TypeDescriptor<CreatedPoolEvent>(
      CreatedPoolEvent.TYPE_QNAME,
    );

    export function type(): TypeDescriptor<CreatedPoolEvent> {
      return TYPE.apply();
    }
  }

  export type CreatedPoolEventInstance =
    TypedEventInstance<CreatedPoolEvent> & {
      data_decoded: CreatedPoolEvent;
      type_arguments: [];
    };

  export interface DepositEvent {
    pool_id: _0x2.object$.ID;
    issuer: string;
    referrer: _0x1.option.Option<string>;
    types: _0x1.ascii.String[];
    deposits: bigint[];
    lp_coins_minted: bigint;
  }

  export namespace DepositEvent {
    export const TYPE_QNAME =
      "0xefe170ec0be4d762196bedecd7a065816576198a6527c99282a2551aaa7da38c::events::DepositEvent";

    const TYPE = new TypeDescriptor<DepositEvent>(DepositEvent.TYPE_QNAME);

    export function type(): TypeDescriptor<DepositEvent> {
      return TYPE.apply();
    }
  }

  export type DepositEventInstance = TypedEventInstance<DepositEvent> & {
    data_decoded: DepositEvent;
    type_arguments: [];
  };

  export interface DepositEventV2 {
    pool_id: _0x2.object$.ID;
    issuer: string;
    referrer: _0x1.option.Option<string>;
    types: _0x1.ascii.String[];
    deposits: bigint[];
    reserves: bigint[];
    lp_coins_minted: bigint;
  }

  export namespace DepositEventV2 {
    export const TYPE_QNAME =
      "0xefe170ec0be4d762196bedecd7a065816576198a6527c99282a2551aaa7da38c::events::DepositEventV2";

    const TYPE = new TypeDescriptor<DepositEventV2>(DepositEventV2.TYPE_QNAME);

    export function type(): TypeDescriptor<DepositEventV2> {
      return TYPE.apply();
    }
  }

  export type DepositEventV2Instance = TypedEventInstance<DepositEventV2> & {
    data_decoded: DepositEventV2;
    type_arguments: [];
  };

  export interface OraclePriceEvent {
    pool_id: _0x2.object$.ID;
    base_type: _0x1.ascii.String;
    quote_type: _0x1.ascii.String;
    oracle_price: bigint;
  }

  export namespace OraclePriceEvent {
    export const TYPE_QNAME =
      "0xefe170ec0be4d762196bedecd7a065816576198a6527c99282a2551aaa7da38c::events::OraclePriceEvent";

    const TYPE = new TypeDescriptor<OraclePriceEvent>(
      OraclePriceEvent.TYPE_QNAME,
    );

    export function type(): TypeDescriptor<OraclePriceEvent> {
      return TYPE.apply();
    }
  }

  export type OraclePriceEventInstance =
    TypedEventInstance<OraclePriceEvent> & {
      data_decoded: OraclePriceEvent;
      type_arguments: [];
    };

  export interface SpotPriceEvent {
    pool_id: _0x2.object$.ID;
    base_type: _0x1.ascii.String;
    quote_type: _0x1.ascii.String;
    spot_price: bigint;
  }

  export namespace SpotPriceEvent {
    export const TYPE_QNAME =
      "0xefe170ec0be4d762196bedecd7a065816576198a6527c99282a2551aaa7da38c::events::SpotPriceEvent";

    const TYPE = new TypeDescriptor<SpotPriceEvent>(SpotPriceEvent.TYPE_QNAME);

    export function type(): TypeDescriptor<SpotPriceEvent> {
      return TYPE.apply();
    }
  }

  export type SpotPriceEventInstance = TypedEventInstance<SpotPriceEvent> & {
    data_decoded: SpotPriceEvent;
    type_arguments: [];
  };

  export interface SwapEvent {
    pool_id: _0x2.object$.ID;
    issuer: string;
    referrer: _0x1.option.Option<string>;
    types_in: _0x1.ascii.String[];
    amounts_in: bigint[];
    types_out: _0x1.ascii.String[];
    amounts_out: bigint[];
  }

  export namespace SwapEvent {
    export const TYPE_QNAME =
      "0xefe170ec0be4d762196bedecd7a065816576198a6527c99282a2551aaa7da38c::events::SwapEvent";

    const TYPE = new TypeDescriptor<SwapEvent>(SwapEvent.TYPE_QNAME);

    export function type(): TypeDescriptor<SwapEvent> {
      return TYPE.apply();
    }
  }

  export type SwapEventInstance = TypedEventInstance<SwapEvent> & {
    data_decoded: SwapEvent;
    type_arguments: [];
  };

  export interface SwapEventV2 {
    pool_id: _0x2.object$.ID;
    issuer: string;
    referrer: _0x1.option.Option<string>;
    types_in: _0x1.ascii.String[];
    amounts_in: bigint[];
    types_out: _0x1.ascii.String[];
    amounts_out: bigint[];
    reserves: bigint[];
  }

  export namespace SwapEventV2 {
    export const TYPE_QNAME =
      "0xefe170ec0be4d762196bedecd7a065816576198a6527c99282a2551aaa7da38c::events::SwapEventV2";

    const TYPE = new TypeDescriptor<SwapEventV2>(SwapEventV2.TYPE_QNAME);

    export function type(): TypeDescriptor<SwapEventV2> {
      return TYPE.apply();
    }
  }

  export type SwapEventV2Instance = TypedEventInstance<SwapEventV2> & {
    data_decoded: SwapEventV2;
    type_arguments: [];
  };

  export interface WithdrawEvent {
    pool_id: _0x2.object$.ID;
    issuer: string;
    referrer: _0x1.option.Option<string>;
    types: _0x1.ascii.String[];
    withdrawn: bigint[];
    lp_coins_burned: bigint;
  }

  export namespace WithdrawEvent {
    export const TYPE_QNAME =
      "0xefe170ec0be4d762196bedecd7a065816576198a6527c99282a2551aaa7da38c::events::WithdrawEvent";

    const TYPE = new TypeDescriptor<WithdrawEvent>(WithdrawEvent.TYPE_QNAME);

    export function type(): TypeDescriptor<WithdrawEvent> {
      return TYPE.apply();
    }
  }

  export type WithdrawEventInstance = TypedEventInstance<WithdrawEvent> & {
    data_decoded: WithdrawEvent;
    type_arguments: [];
  };

  export interface WithdrawEventV2 {
    pool_id: _0x2.object$.ID;
    issuer: string;
    referrer: _0x1.option.Option<string>;
    types: _0x1.ascii.String[];
    withdrawn: bigint[];
    reserves: bigint[];
    lp_coins_burned: bigint;
  }

  export namespace WithdrawEventV2 {
    export const TYPE_QNAME =
      "0xefe170ec0be4d762196bedecd7a065816576198a6527c99282a2551aaa7da38c::events::WithdrawEventV2";

    const TYPE = new TypeDescriptor<WithdrawEventV2>(
      WithdrawEventV2.TYPE_QNAME,
    );

    export function type(): TypeDescriptor<WithdrawEventV2> {
      return TYPE.apply();
    }
  }

  export type WithdrawEventV2Instance = TypedEventInstance<WithdrawEventV2> & {
    data_decoded: WithdrawEventV2;
    type_arguments: [];
  };
// ...
}